requirements.txt
Pillow>=9.1.1
tkinterdnd2>=0.3.0
